0.9.0 (filter, deprecated)
"filter" (in "cmap") has been renamed as "range". Also, "range" now
accepts multipliers in "pyram" and "rank". Previously, x-axis in "pyram"
and "rank" was increased by 2% (1.02) and this couldn't be changed. Now
this value is set in the setup file.

Consequently, the setup file has changed.

"filter" is deprecated: support for filter will be removed in next
version.